refactor: remove dead code and resolve structural issues
