+++ ark-taco-shop-api/.prettier.config.js
@@ -1,7 +1,7 @@
 // prettier.config.js or .prettierrc.js
 module.exports = {
-  trailingComma: "es5",
-  tabWidth: 4,
-  semi: false,
-  singleQuote: true
-}
+    trailingComma: "es5",
+    tabWidth: 4,
+    semi: false,
+    singleQuote: true,
+};
