+++ ark-taco-shop-api/jest.config.js
@@ -1,22 +1,14 @@
-'use strict'
+"use strict";
 
 module.exports = {
-  testEnvironment: 'node',
-  bail: false,
-  verbose: true,
-  testMatch: [
-    '**/__tests__/**/*.test.js'
-  ],
-  moduleFileExtensions: [
-    'js',
-    'json'
-  ],
-  collectCoverage: false,
-  coverageDirectory: '<rootDir>/.coverage',
-  collectCoverageFrom: [
-    'lib/**/*.js',
-    '!**/node_modules/**'
-  ],
-  watchman: false,
-  setupTestFrameworkScriptFile: 'jest-extended'
-}
+    testEnvironment: "node",
+    bail: false,
+    verbose: true,
+    testMatch: ["**/__tests__/**/*.test.js"],
+    moduleFileExtensions: ["js", "json"],
+    collectCoverage: false,
+    coverageDirectory: "<rootDir>/.coverage",
+    collectCoverageFrom: ["lib/**/*.js", "!**/node_modules/**"],
+    watchman: false,
+    setupTestFrameworkScriptFile: "jest-extended",
+};
