+++ ark-taco-shop-api/jsdoc.json
@@ -1,24 +1,22 @@
 {
-  "tags": {
-    "allowUnknownTags": false
-  },
-  "source": {
-    "include": "./lib",
-    "includePattern": ".js$",
-    "excludePattern": "(node_modules/|docs)"
-  },
-  "plugins": [
-    "plugins/markdown"
-  ],
-  "opts": {
-    "template": "../../node_modules/docdash",
-    "encoding": "utf8",
-    "destination": "docs/",
-    "recurse": true,
-    "verbose": true
-  },
-  "templates": {
-    "cleverLinks": false,
-    "monospaceLinks": false
-  }
+    "tags": {
+        "allowUnknownTags": false
+    },
+    "source": {
+        "include": "./lib",
+        "includePattern": ".js$",
+        "excludePattern": "(node_modules/|docs)"
+    },
+    "plugins": ["plugins/markdown"],
+    "opts": {
+        "template": "../../node_modules/docdash",
+        "encoding": "utf8",
+        "destination": "docs/",
+        "recurse": true,
+        "verbose": true
+    },
+    "templates": {
+        "cleverLinks": false,
+        "monospaceLinks": false
+    }
 }
