--- ark-taco-shop-api/lib/server/public/javascripts/smoothScroll.js
@@ -1,3 +1,0 @@
-$(document).ready(function () {
-  SmoothScroll('a[href*="#"]');
-});
