--- ark-taco-shop-api/lib/types/index.d.ts
@@ -1,7 +1,0 @@
-import { DataTypeAbstract, DefineAttributeColumnOptions } from "sequelize";
-
-declare global {
-  type SequelizeAttributes<T extends { [key: string]: any }> = {
-    [P in keyof T]: string | DataTypeAbstract | DefineAttributeColumnOptions
-  };
-}
