+++ ark-taco-shop-api/package.json
@@ -1,68 +1,60 @@
 {
-  "private": true,
-  "name": "@mlh/ark-taco-shop-api",
-  "description": "Provides an API for taco orders on the Ark Core",
-  "version": "0.0.1",
-  "contributors": [
-    "Halisson Bruno <halissonvit@gmail.com>"
-  ],
-  "license": "MIT",
-  "main": "dist/index",
-  "types": "dist/index",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "start": "tsc && node ./bin/www",
-    "compile": "./node_modules/typescript/bin/tsc",
-    "build": "yarn clean && yarn build:prepare && yarn compile && yarn build:assets",
-    "build:assets": "cp -rf ./lib/server/public ./dist/server/",
-    "build:prepare": "mkdir dist",
-    "build:watch": "yarn clean && yarn build && yarn compile -w",
-    "clean": "del dist",
-    "lint": "../../node_modules/tslint/bin/tslint -c ../../tslint.json 'src/**/*.ts' '__tests__/**/*.ts' --fix"
-  },
-  "dependencies": {
-    "@arkecosystem/core-container": "^0.1.1",
-    "@arkecosystem/core-interfaces": "^2.1.1",
-    "axios": "^0.18.0",
-    "boom": "^7.2.0",
-    "del-cli": "^1.1.0",
-    "fs-extra": "^7.0.0",
-    "h2o2": "^8.1.2",
-    "hapi": "^17.5.0",
-    "hapi-pagination": "^2.0.0",
-    "inert": "^5.1.1",
-    "joi": "^13.3.0",
-    "micromatch": "^3.1.10",
-    "pg": "^7.8.2",
-    "request-ip": "^2.0.2",
-    "sequelize": "^4.37.8",
-    "sqlite3": "^4.0.2",
-    "typescript": "^3.3.3333",
-    "umzug": "^2.1.0",
-    "wreck": "^14.1.0"
-  },
-  "devDependencies": {
-    "@sindresorhus/tsconfig": "^0.2.1",
-    "@types/boom": "^7.2.1",
-    "@types/hapi": "^18.0.1",
-    "@types/node": "^11.10.4",
-    "@types/sequelize": "^4.27.39",
-    "@types/wreck": "^14.0.0",
-    "depcheck": "^0.6.9",
-    "eslint": "^4.19.1",
-    "eslint-config-standard": "^11.0.0",
-    "eslint-plugin-import": "^2.12.0",
-    "eslint-plugin-node": "^6.0.1",
-    "eslint-plugin-promise": "^3.8.0",
-    "eslint-plugin-standard": "^3.1.0",
-    "jest": "^23.1.0",
-    "jest-extended": "^0.7.2",
-    "jsdoc": "^3.5.5",
-    "prettier": "1.14.3",
-    "regenerator-runtime": "^0.11.1",
-    "standard": "^11.0.1",
-    "tsc": "^1.20150623.0"
-  }
+    "private": true,
+    "name": "@mlh/ark-taco-shop-api",
+    "description": "Provides an API for taco orders on the Ark Core",
+    "version": "0.0.1",
+    "contributors": [
+        "Halisson Bruno <halissonvit@gmail.com>"
+    ],
+    "license": "MIT",
+    "main": "dist/index",
+    "types": "dist/index",
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "start": "tsc && node ./bin/www",
+        "compile": "tsc",
+        "build": "yarn clean && yarn build:prepare && yarn compile && yarn build:assets",
+        "build:assets": "cp -rf ./src/server/public ./dist/server/",
+        "build:prepare": "mkdir dist",
+        "build:watch": "yarn clean && yarn build && yarn compile -w",
+        "clean": "del dist",
+        "format": "tslint -c ./tslint.json 'src/**/*.ts' --fix && prettier --write \"./*.{ts,js,json,md}\" \"./src/**/*.{ts,js,json,md}\""
+    },
+    "dependencies": {
+        "@arkecosystem/core-container": "^2.3.0-next.4",
+        "@arkecosystem/core-http-utils": "^2.3.0-next.4",
+        "@arkecosystem/core-interfaces": "^2.3.0-next.4",
+        "axios": "^0.18.0",
+        "boom": "^7.2.0",
+        "del-cli": "^1.1.0",
+        "fs-extra": "^7.0.0",
+        "h2o2": "^8.1.2",
+        "hapi": "^17.5.0",
+        "hapi-pagination": "^2.0.0",
+        "inert": "^5.1.1",
+        "joi": "^13.3.0",
+        "micromatch": "^3.1.10",
+        "pg": "^7.8.2",
+        "request-ip": "^2.0.2",
+        "sequelize": "^4.37.8",
+        "sqlite3": "^4.0.2",
+        "typescript": "^3.3.3333",
+        "umzug": "^2.1.0",
+        "wreck": "^14.1.0"
+    },
+    "devDependencies": {
+        "@sindresorhus/tsconfig": "^0.2.1",
+        "@types/express": "^4.16.1",
+        "@types/http-proxy-middleware": "^0.19.2",
+        "@types/node": "^11.10.5",
+        "cross-env": "^5.2.0",
+        "jest": "^23.1.0",
+        "jest-extended": "^0.7.2",
+        "prettier": "^1.16.4",
+        "tslint": "^5.14.0",
+        "tslint-config-prettier": "^1.18.0",
+        "typescript": "^3.3.3333"
+    }
 }
