+++ ark-taco-shop-api/src/server/public/images/ark-logo.png
