+++ ark-taco-shop-api/src/server/public/inventory.html
