+++ ark-taco-shop-api/src/server/public/javascripts/smoothScroll.js
@@ -1,0 +1,3 @@
+$(document).ready(function() {
+    SmoothScroll('a[href*="#"]');
+});
