+++ ark-taco-shop-api/src/server/public/stylesheets/style.css
