+++ ark-taco-shop-api/src/types/index.d.ts
@@ -1,0 +1,7 @@
+import { DataTypeAbstract, DefineAttributeColumnOptions } from "sequelize";
+
+declare global {
+    type SequelizeAttributes<T extends { [key: string]: any }> = {
+        [P in keyof T]: string | DataTypeAbstract | DefineAttributeColumnOptions
+    };
+}
